Refactor and split code for better modularity
